Adding reverse proxying for files to the frontend.
The users should be speaking either to the frontend or something else to
get the static files and not directly to the backend.

The long term answer for this is to have a shared filesystem so we can
directly serve up the same files from the frontend to the users that the
drupal backend stores but for now this will do.