Adding in scale commands to dev deploy.
We have found that one container for frontend and backend is not enough
so by default scaling to three containers for the frontend and backend.

This requires we not only scale the stack but add a sleep to wait for
the stack to finish provisioning. This is a quick hack and we should
wait on the stack provisioning properly before trying to test it.